@@ -133,7 +133,7 @@
     "madge:circular": "node_modules/.bin/madge ./src --circular"
   },
   "engines": {
-    "node": ">=12.22.10 <19"
+    "node": ">=14.21.0 <17 || >=18 <19"
   },
   "bin": {
     "parse-server": "bin/parse-server"
