@@ -1,13 +1,13 @@
-import redis from 'redis';
+import { createClient } from 'redis';
 
 function createPublisher({ redisURL, redisOptions = {} }): any {
   redisOptions.no_ready_check = true;
-  return redis.createClient(redisURL, redisOptions);
+  return createClient(redisURL, redisOptions);
 }
 
 function createSubscriber({ redisURL, redisOptions = {} }): any {
   redisOptions.no_ready_check = true;
-  return redis.createClient(redisURL, redisOptions);
+  return createClient(redisURL, redisOptions);
 }
 
 const RedisPubSub = {
